Removed bouncer auth references.
